Refacto UI : remove AppService
