+++ ui/main/src/app/business/store/router.store.ts
@@ -14,6 +14,19 @@ import {ReplaySubject, Observable} from 'rxjs';
     providedIn: 'root'
 })
 export class RouterStore {
+
+    private pageConf = new Map([
+        ['feed', PageType.FEED],
+        ['archives', PageType.ARCHIVE],
+        ['businessconfigparty', PageType.THIRPARTY],
+        ['setting', PageType.SETTING],
+        ['about', PageType.ABOUT],
+        ['calendar', PageType.CALENDAR],
+        ['monitoring', PageType.MONITORING],
+        ['usercard', PageType.USERCARD]
+    ]);
+
+
     private currentRouteEvent = new ReplaySubject<string>(1);
     private currentRoute = 'feed';
 
@@ -34,4 +47,24 @@ export class RouterStore {
     private isRouteUrlAfterImplicitAuth(route: string): boolean {
         return route.startsWith('state');
     }
+
+
+    public getCurrentPageType(): PageType {
+        const pageName = this.getCurrentRoute()
+        const currentPageType = this.pageConf.get(pageName);
+        return !!currentPageType ? currentPageType : PageType.UNKNOWN;
+    }
+
+}
+
+export enum PageType {
+    UNKNOWN,
+    FEED,
+    ARCHIVE,
+    THIRPARTY,
+    SETTING,
+    ABOUT,
+    CALENDAR,
+    MONITORING,
+    USERCARD
 }
