+++ ui/main/src/app/modules/feed/components/card-list/card-list.component.ts
@@ -15,13 +15,13 @@ import {NgbModal} from '@ng-bootstrap/ng-bootstrap';
 import {ConfigService} from 'app/business/services/config.service';
 import {MessageLevel} from '@ofModel/message.model';
 import {ProcessesService} from 'app/business/services/processes.service';
-import {AppService} from '@ofServices/app.service';
 import {AcknowledgeService} from 'app/business/services/acknowledge.service';
 import {UserService} from 'app/business/services/user.service';
 import {UserWithPerimeters} from '@ofModel/userWithPerimeters.model';
 import {EntitiesService} from 'app/business/services/entities.service';
 import {GroupedCardsService} from 'app/business/services/grouped-cards.service';
 import {AlertMessageService} from 'app/business/services/alert-message.service';
+import {Router} from '@angular/router';
 
 @Component({
     selector: 'of-card-list',
@@ -46,10 +46,10 @@ export class CardListComponent implements AfterViewChecked, OnInit {
         private processesService: ProcessesService,
         private acknowledgeService: AcknowledgeService,
         private userService: UserService,
-        private _appService: AppService,
         private entitiesService: EntitiesService,
         private groupedCardsService: GroupedCardsService,
-        private alertMessageService: AlertMessageService
+        private alertMessageService: AlertMessageService,
+        private router: Router
     ) {
         this.currentUserWithPerimeters = this.userService.getCurrentUserWithPerimeters();
     }
@@ -124,7 +124,7 @@ export class CardListComponent implements AfterViewChecked, OnInit {
     confirmAckAllCards() {
         this.modalRef.close();
         this.acknowledgeAllVisibleCardsInTheFeed();
-        this._appService.closeDetails();
+        this.router.navigate(['/feed']);
     }
 
     declineAckAllCards(): void {
